Don't rely on hash ordering [1].
The profile was previously contained in a hash, mapping the name to an array
of its devices. Obviously, there is no guarantee that we iterated in the same
way that we added. In other words, what happened was not a top-to-bottom
parsing of the configuration file, but more like a random parsing.

[1] http://qtmemes.tumblr.com/image/20282967408