Adding option for user defined rules to profiles.
An :only_if or :not_if option may be specified with a profile. The value
should correspond to a method defined within the configuration file and
should return whether or not the profile can be invoked.

The following code shows an example of how this is used.

  def laptop_lid_open?
    File.read('/proc/acpi/button/lid/LID/state').match(/open/)
  end

  profile "docked", :only_if => :laptop_lid_open? do
    device "LVDS1"
    device "VGA1", :right_of => "LVDS1"
  end

This fixes issue #1.